requirements-dev.txt
# Development dependencies for Banking Transactions API
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Linting and type checking
flake8>=6.1.0
mypy>=1.7.0
pandas-stubs>=2.1.0
